+++ Cmdlets/Public/Get-TeamViewerAddressBook.ps1
@@ -1,0 +1,28 @@
+﻿function Get-TeamViewerAddressBook {
+    [CmdletBinding()]
+
+    [OutputType([psobject])]
+
+    param(
+        [Parameter(Mandatory = $true)]
+        [securestring]
+        $APIToken
+    )
+
+    $Resource_Uri = "$(Get-TeamViewerAPIUri)/companyaddressbook"
+    $Parameters = @{}
+
+    do {
+        $Response = Invoke-TeamViewerRestMethod `
+            -APIToken $APIToken `
+            -Uri $Resource_Uri `
+            -Method Get `
+            -Body $Parameters `
+            -WriteErrorTo $PSCmdlet `
+            -ErrorAction Stop
+
+        Write-Output $Response
+
+        $Parameters.ct = $Response.continuation_token
+    } while ($Parameters.ct)
+}